Evaluate the segmentation against ground truth
The code will take the intermediate results generated by the worker and
caclulate the rand and voi score of the whole segmentation. Only tested
with small cutouts (10kx10kx2k volume). The code will also collect the
pairs of segments differ significantly and generate a json file for
later reviewing