Adjust the heartbeat setting of rabbitmq
The current heartbeat does not work when the workers are busy, my guess
is the amqp connections is created in the main thread, so even when the
heartbeat_check function is called from a thread, it still need to run
in the main thread, causing the heartbeat to miss. Somehow the code
works if instead of doing actual task I use a simple sleep.