Change the scaling behavior to match the default setup
By default mip3 is the batch mip and mip5 is the high mip, the atomic
workers should be shutdown completely when mip4 finishes, but the old
behavior will reduce the number of workers to match the number of tasks
at mip5, which is not necessary.

When using different batch mip and high mip setup, I hope we can now
rely on the cluser_manager dag to scale up/dowm the cluster accordingly.