Chunkflow dag for the bot
The current approach create the same number of chunkflow operator as the
number of chunks, but there is no binding relation between them, the
operator can process as many chunks as it can receive from the task
queue. The rest of the operators will terminate themselves after they
get task queue empty message 5 times.

It is possible to do this without the dag by directly creating gpu
instances and running chunkflow from the startup scripts, but the
current approach allows us to view the error messages in airflow web
interface