Ugly hack to stop igneous tasks
Airflow does not provide a method to actually stop a running task, even
if we mark the task as success the worker continues to run it until it
actually finishes. This is a problem for downsample and meshing tasks
if the user cancels the run in the middle. The hack uses the "run_token"
variable, and stop the task if the token is an empty string. On the
other side the bot will set the token to empty once it cancels the
current run