Allow custom workers to return messages
When the process_task function return a dict containing a 'ret' entry,
the value will be json dumped and send to a returning queue, the server
will download the messages and print them to the slack channel. Useful
for notification in sparse scenarios. However it will flood the slack
channel if every worker send some message for every task.

Maybe a useful feature in the future to allow user supply their own
function to process the returning message.