Fix task retry in chunkflow dags
Chunkflow dags set the retry policy using default parameters giving to
the dag constructors, so there is no explicit default_args given when
creating the operators. Fallback to the default_args in the dag if no
default_args is passed to worker_op. Default_args will be an empty dict
by default so there is no need to add another layer of fallback.