Support multiple types of custom tasks in one task script
In addition to the original interface where the script returns the tasks
in a list, allow the script to provide a "task_generator" entry in the
dict to yield different types of tasks in each iteration. The
process_task function is responsible for handle different type of tasks
properly. For example the tasks may have a "custom_task_type" entry in
the payload and process_task dispatches the tasks to their own handler
based on this flag. A sketch of the workflow looks like this:

def submit_tasks_type1(...):
    tasks = []
    for ...
        task['custom_task_type'] = 'type1'
        ...
        tasks.append(task)
    return tasks

def submit_tasks_type2(...):
    tasks = []
    for ...
        task['custom_task_type'] = 'type2'
        ...
        tasks.append(task)
    return tasks

def task_generator(...):
    yield submit_tasks_type1(...)
    yield submit_tasks_type2(...)
    ...

def submit_tasks():
    return {'task_generator': task_generator(...)}

def process_task(msg):
    t = json.loads(msg)
    if t['custom_task_type'] == 'type1':
        ...
    elif t['custom_task_type'] == 'type1':
        ...
    ...