Add an "INHERIT_PARAMETERS" option
In some cases using the first set of parameters in a batch of runs as
templates causes conflicts. For example when we want to compare
inference output between pytorch models and onnx models. The easiest
solution is an "INHERIT_PARAMETERS" so we can explicitly disable the
behavior. One do have to repeat all the common parameters in this case,
but can be solved by using a python script to generate the parameters.