Change slack_message interface
The old interface allowed sending messages to arbitrary channels, but
the only use case is to send messages to the seuron-alerts channel. It
makes more send to add a notification flag, and send messages to
seuron-alerts when the flag is true. Otherwise just send the message to
the thread user started.