Do not use setdefault in the dag
Due to recent changes in airflow

https://github.com/apache/airflow/pull/16404

get variable will never raise an exception, it will return a None
instead. This will cause setdefault call to reset everything to default
if the database is not available when it try to fetch the value. For now
we set the default value once during the initialization and let the dag
fail if it gets a None parameter.