Replace create_run_token call with cancelable attribute in the decorator
The token are created after calling the command function, so even if the
command failed immediately, a token will still be create. It does not
really hurt any functionality but feels awkward when it happens, we can
have the command function to return a boolean to indicate if a token
should be create.

In principle we should hide the cancel run operation in the bot as well
since it is universal and closely coupled to the run token system, but
right now it have to live in slack_bot.py in order to interact with the
worker queue.