Only reset the igneous/custom cluster for long jobs
Keep the cluster running if the tasks finish within 10 minutes, this
mainly helps the segmentation jobs where several types of igneous tasks
are executed sequentially. If the cutout is smaller, there is no reason
to reset the instances over and over again. To make sure all instances
are shutdown in the end, add a scaling operation at the end of the
igneous/custom dags.