Allow docker compose deployment to process local dataset
If vender is set to LocalDockerCompose, when starting a worker
container, the operator will inspect itself and find out the extra
mounted volumes except /tmp and /var/run/docker.sock. These volumes
will be mounted to the worker container as well. The operator will also
connect the worker container to the network created by docker compose,
so it can use all the services in the deployment