Better estimation of the instance needed
We allow multiple workers running on one instance, it was inaccurate and
wasteful to scale the instances to match the number of tasks.

There are two cases the current code does not cover, the first is the
composite workers, I simply assume one instance for each worker. This is
an overestimation. The other case is the igneous/custom tasks specifying
its own concurrency, because it cannot be higher than the maximal
concurrency of the setup, it is an underestimation.