Fail the sanity check if bbox does not align with existing ws layer
Because the watershed layer are chunked, if the user want rerun
agglomeration with an existing watershed layer, the bbox must be
aligned in order to use the chunkmap data to reconstruct the unchunked
supervoxels. In principle we only need the minpt to align, but it is
safer to request the entire bbox be aligned.