Add an optional easyseg worker instance
This worker instance runs the workers required for all tasks in the easyseg
process. The necessary tools are chunkflow, abiss, and igneous. The setup is
done during deployment time, and the instance should only be activated right
before starting the easyseg task. Once the process is completed, the instance
should be shut down.